requirements.txt
numpy==1.23.0
scikit-learn
scipy
pandas
matplotlib
nltk
spacy
tqdm
unidecode
fvcore
lxml
psutil
Pillow==8.4.0
requests
beautifulsoup4
tldextract
gdown
selenium
selenium-wire
helium
webdriver-manager
flask
flask-cors
google-cloud-vision
googletrans
editdistance
cryptography==38.0.4
httpcore==0.15.0
h11
h2
blinker==1.7.0
hyperframe
pycocotools
opencv-python
opencv-contrib-python
